requirements.txt
streamlit
pandas
matplotlib
plotly
requests
openpyxl 
python-dotenv 